Fix deleting countries

+++ 02_Domain/CountryDAO.cs
@@ -4,17 +4,20 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using static System.ComponentModel.Design.ObjectSelectorEditor;
+using System.Windows.Forms;
 
 namespace BazyDanychBadminton._02_Domain
 {
-    public class CountryDAO {
+    public class CountryDAO
+    {
         public List<Country> ReadAll()
         {
             List<Country> result = new List<Country>();
             string sql = "SELECT * FROM Countries ORDER BY idCountry;";
             List<string[]> table = DBBroker.getInstance().Read(sql);
-            foreach (string[] row in table) 
-            { 
+            foreach (string[] row in table)
+            {
                 Country c = new Country();
                 c.IdCountry = row[0];
                 c.CountryName = row[1];
@@ -42,6 +45,24 @@ namespace BazyDanychBadminton._02_Domain
                 c.IdCountry = row[0];
             }
         }
+        public bool CanDelete(Country c)
+        {
+            string sql = $@"
+            SELECT
+            (SELECT COUNT(*) FROM players WHERE plaCountry = '{c.IdCountry}') AS playersCount,
+            (SELECT COUNT(*) FROM tournaments WHERE touCountry = '{c.IdCountry}') AS tournamentsCount;";
+
+
+            List<string[]> result = DBBroker.getInstance().Read(sql);
+            if (result.Count > 0)
+            {
+                string[] row = result[0];
+                int playersCount = Convert.ToInt32(row[0]);
+                int tournamentsCount = Convert.ToInt32(row[1]);
+                return playersCount == 0 && tournamentsCount == 0;
+            }
+            return false;
+        }
         public int Insert(Country c)
         {
             string sql = "INSERT INTO Countries VALUES ('" + c.IdCountry + "', '" + c.CountryName + "');";
@@ -55,6 +76,10 @@ namespace BazyDanychBadminton._02_Domain
         }
         public int Delete(Country c)
         {
+            if (!CanDelete(c))
+            {
+                return -1; // Cannot delete country, it has players
+            }
             string sql = "DELETE FROM Countries WHERE idCountry='" + c.IdCountry + "';";
             return DBBroker.getInstance().Change(sql);
         }
